Resolve "Couple Whisper with the game controller" and "Couple ALIZE with the game controller"

See merge request iglesias/2022-sb1-ga-gytmy!87

+++ src/main/java/com/gytmy/maze/controller/MazeControllerImplementation.java
@@ -15,8 +15,12 @@ import com.gytmy.maze.model.score.ScoreCalculator;
 import com.gytmy.maze.model.score.ScoreType;
 import com.gytmy.maze.view.game.MazeView;
 import com.gytmy.maze.view.game.MazeViewFactory;
+import com.gytmy.sound.AudioFileManager;
+import com.gytmy.sound.AudioRecognitionResult;
 import com.gytmy.sound.AudioRecorder;
 import com.gytmy.sound.RecordObserver;
+import com.gytmy.sound.User;
+import com.gytmy.sound.AlizeRecognitionResultParser.AlizeRecognitionResult;
 import com.gytmy.sound.whisper.Whisper;
 import com.gytmy.sound.whisper.Whisper.Model;
 import com.gytmy.utils.Coordinates;
@@ -29,8 +33,10 @@ public class MazeControllerImplementation implements MazeController, RecordObser
     private MazeView view;
     private JFrame frame;
     private boolean hasCountdownEnded = false;
+    private boolean compareWithWhisper = true; // true if the compare function should use Whisper
+
     private static String FILE_NAME = "currentGameAudio";
-    private static String AUDIO_GAME_PATH = "src/resources/audioFiles/client/audio/"+ FILE_NAME + ".wav";
+    private static String AUDIO_GAME_PATH = "src/resources/audioFiles/client/audio/" + FILE_NAME + ".wav";
     private static String JSON_OUTPUT_PATH = "src/resources/audioFiles/client/audio/model/json/";
 
     private MovementControllerType selectedMovementControllerType = MovementControllerType.KEYBOARD;
@@ -107,20 +113,68 @@ public class MazeControllerImplementation implements MazeController, RecordObser
         }, "Record Audio In Game");
     }
 
+    /**
+     * Compare the audio with model and move the player in consequence
+     * 
+     * (the direction is determined with whisper or with alize, depends on
+     * compareWithWhisper)
+     */
     private void compareAudioWithModel() {
+        AlizeRecognitionResult result = AudioRecognitionResult.getRecognitionResult();
 
-        CompletableFuture<String> futureCommand = whisper.ask(AUDIO_GAME_PATH, FILE_NAME, JSON_OUTPUT_PATH);
+        if (result == null) {
+            return;
+        }
 
-        futureCommand.thenAccept(recognizedCommand -> {
+        User recognizedUser = AudioFileManager.getUser(result.getName());
 
-            // TODO : @gdudilli - Ici pour recuperer la commande reconnue par Whisper 
-            System.out.println("\nrecognizedCommand: " + recognizedCommand);
-            System.out.println("-----------");
+        if (recognizedUser == null) {
+            return;
+        }
 
-            new File(AUDIO_GAME_PATH).delete();
-            new File(JSON_OUTPUT_PATH + FILE_NAME + ".json").delete();
-        });
-        
+        if (!compareWithWhisper) {
+            movePlayerWithCompareResult(recognizedUser, result.getWord());
+        } else {
+            CompletableFuture<String> futureCommand = whisper.ask(AUDIO_GAME_PATH, FILE_NAME, JSON_OUTPUT_PATH);
+
+            futureCommand.thenAccept(recognizedCommand -> {
+
+                System.out.println("------------------------------------");
+                System.out.println("recognizedCommand: " + recognizedCommand);
+                System.out.println("recognizedUser: " + recognizedUser.getUserName());
+                System.out.println("------------------------------------");
+
+                movePlayerWithCompareResult(recognizedUser, recognizedCommand);
+
+                new File(JSON_OUTPUT_PATH + FILE_NAME + ".json").delete();
+            });
+        }
+
+        new File(AUDIO_GAME_PATH).delete();
+    }
+
+    /**
+     * Move the player if its name is recognized, in the direction of
+     * nameOfDirection
+     * 
+     * @param recognizedUser
+     * @param directionName
+     */
+    private void movePlayerWithCompareResult(User recognizedUser, String directionName) {
+        Player[] players = getPlayers();
+
+        for (Player player : players) {
+
+            if (recognizedUser.getUserName().equals(player.getName())) {
+
+                Direction direction = Direction.stringToDirection(directionName);
+
+                if (direction != null) {
+                    movePlayer(player, direction);
+                }
+                return;
+            }
+        }
     }
 
     @Override
